IMU changed, changes to Drivtrain encoder Using ADIS16470 now. Using ControlMode.Position for PID encoder controlled Talon.

+++ 2020-competition-code/src/main/java/frc/robot/commandGroups/DriveToPoint.java
@@ -24,7 +24,7 @@ public class DriveToPoint extends SequentialCommandGroup {
     super();
     
     double angle = Math.atan(Y/X);
-    double hyp = Math.sqrt((X*X) + (Y*Y));
+    int hyp = (int)(Math.sqrt((X*X) + (Y*Y)));
 
     addCommands(
